requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
django-urls==1.1.3
Django-Verify-Email==1.0.7
gunicorn==20.1.0
numpy==1.21.0
Pillow==8.2.0
psycopg2==2.9.1
pycodestyle==2.7.0
python-decouple==3.4
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
whitenoise==5.2.0
xlwt==1.3.0
